chore: split long running tests
