+++ .circleci/config.base.yml
@@ -756,7 +756,7 @@ workflows:
               - cleanup-resources
               - e2e-test-context
           requires:
-            - build_pkg_binaries
+            - build
           filters:
             branches:
               only:
@@ -775,7 +775,7 @@ workflows:
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
           requires:
-            - cleanup_resources
+            - build_pkg_binaries
       - amplify_migration_tests_v6:
           context:
             - e2e-auth-credentials
@@ -788,7 +788,7 @@ workflows:
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
           requires:
-            - cleanup_resources
+            - build_pkg_binaries
           # post-steps:
           #   - run: *clean_e2e_resources
       - amplify_migration_tests_v5:
@@ -803,7 +803,7 @@ workflows:
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
           requires:
-            - cleanup_resources
+            - build_pkg_binaries
           # post-steps:
           #   - run: *clean_e2e_resources
       - amplify_migration_tests_non_multi_env_layers:
@@ -818,7 +818,7 @@ workflows:
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
           requires:
-            - cleanup_resources
+            - build_pkg_binaries
           # post-steps:
           #   - run: *clean_e2e_resources
       - amplify_migration_tests_multi_env_layers:
@@ -833,7 +833,7 @@ workflows:
                 - /tagged-release\/.*/
                 - /run-e2e\/.*/
           requires:
-            - cleanup_resources
+            - build_pkg_binaries
           # post-steps:
           #   - run: *clean_e2e_resources
       - amplify_console_integration_tests:
