+++ .eslint-dictionary.json
@@ -25,8 +25,8 @@
   "codebase",
   "codegen",
   "codepipeline",
-  "columnify",
   "cognito",
+  "columnify",
   "cors",
   "createCipheriv",
   "creds",
@@ -48,6 +48,7 @@
   "frontend",
   "func",
   "generatable",
+  "geofence",
   "getters",
   "gitignore",
   "globals",
